Move app page to `pages`
