+++ src/widgets/index.ts
@@ -1,4 +1,3 @@
-export { Layout } from "./layout/layout.jsx";
 export { Column } from "./column/column.jsx";
 export { Menu } from "./menu/menu.jsx";
 export { Builder } from "./builder/builder.jsx";
