--- src/widgets/layout/index.ts
@@ -1,1 +1,0 @@
-export { Layout } from "./layout";
